Improve display/hiding unlink all button
