+++ pronto/static/js/entry/references.js
@@ -1,4 +1,5 @@
 import * as modals from "../ui/modals.js";
+import { toggleErrorMessage } from "../ui/utils.js";
 
 export function refresh(accession) {
     return fetch(`/api/entry/${accession}/references/`)
@@ -7,8 +8,36 @@ export function refresh(accession) {
             const elem = document.querySelector('#supp-references .content');
             if (!references.length) {
                 elem.innerHTML = '<p>This entry has no additional references.</p>';
+                document.querySelectorAll('.supp-refs-unlink').forEach(e => e.remove())
                 return;
             }
+            
+            // Unlink all references
+            document.querySelector('.supp-refs-actions').innerHTML += `
+                <div class="supp-refs-unlink ui horizontal divider">&nbsp;</div>
+                <button id="unlink-all-btn" class="supp-refs-unlink ui red button">Unlink all</button>
+            `
+
+            document.querySelector('#unlink-all-btn').addEventListener('click', e => {
+                const errMsg = document.querySelector('#edit-entry .ui.message');
+                toggleErrorMessage(errMsg, null);
+                modals.ask(
+                    'Unlink supplementary references',
+                    `Do you want to unlink all the supplementary references?`,
+                    'Unlink all',
+                    () => {
+                        let url = `/api/entry/${accession}/references/`;
+                        fetch(url, { method: 'DELETE' })
+                            .then(response => response.json())
+                            .then(result => {
+                                if (result.status) {
+                                    refresh(accession).then(() => { $('.ui.sticky').sticky(); });
+                                } else
+                                    toggleErrorMessage(errMsg, result.error);
+                            });
+                    }
+                );
+            })
 
             let html = '<p>The following publications were not referred to in the description, but provide useful additional information.</p><ul class="ui list">';
             references.sort((a, b) => a.year - b.year);
@@ -48,7 +77,7 @@ export function refresh(accession) {
                         'This reference will not be associated to this entry anymore.',
                         'Delete',
                         () => {
-                            fetch(`/api/entry/${accession}/reference/${pubId}/`, {method: 'DELETE'})
+                            fetch(`/api/entry/${accession}/reference/${pubId}/`, { method: 'DELETE' })
                                 .then(response => response.json())
                                 .then(result => {
                                     if (result.status)
