+++ pronto/templates/entry.html
@@ -321,11 +321,9 @@
                         <div class="ui form">
                             <div class="field">
                                 <label>Add reference</label>
-                                <div class="ui action input">
+                                <div class="ui action input supp-refs-actions">
                                     <input placeholder="PubMed ID (e.g. 19913036)" type="text" name="pmid">
                                     <button class="ui primary submit button">Add</button>
-                                    <div class="ui horizontal divider">&nbsp;</div>
-                                    <button id="unlink-all" class="ui red button">Unlink all</button>
                                 </div>
                             </div>
                         </div>
